requirements.txt
djangorestframework_simplejwt